remove swift3.0 support
