--- actionRuntimes/swift3Action/Dockerfile
@@ -1,1 +1,0 @@
-FROM openwhisk/swift3action:1.0.0
