--- actionRuntimes/swift3Action/build.gradle
@@ -1,3 +1,0 @@
-ext.dockerImageName = 'swift3action'
-apply from: '../../gradle/docker.gradle'
-
