+++ core/controller/src/main/resources/apiv1swagger.json
@@ -1558,7 +1558,6 @@
                         "nodejs:8",
                         "python:2",
                         "python:3",
-                        "swift:3",
                         "swift:3.1.1",
                         "java",
                         "blackbox"
