+++ settings.gradle
@@ -7,7 +7,6 @@ include 'actionRuntimes:nodejs8Action'
 include 'actionRuntimes:actionProxy'
 include 'actionRuntimes:pythonAction'
 include 'actionRuntimes:python2Action'
-include 'actionRuntimes:swift3Action'
 include 'actionRuntimes:swift3.1.1Action'
 include 'actionRuntimes:javaAction'
 include 'actionRuntimes:php7.1Action'
