+++ tests/build.gradle
@@ -44,7 +44,6 @@ test.dependsOn([
     ':actionRuntimes:pythonAction:distDocker',
     ':actionRuntimes:python2Action:distDocker',
     ':actionRuntimes:javaAction:distDocker',
-    ':actionRuntimes:swift3Action:distDocker',
     ':actionRuntimes:swift3.1.1Action:distDocker',
     ':sdk:docker:distDocker',
     ':tests:dat:blackbox:badaction:distDocker',
