+++ tests/src/test/scala/system/basic/WskBasicSwift3Tests.scala
@@ -37,7 +37,7 @@ abstract class WskBasicSwift3Tests extends TestHelpers with WskTestHelpers with 
   implicit val wskprops: common.WskProps = WskProps()
   val wsk: BaseWsk
   val defaultAction: Some[String] = Some(TestUtils.getTestActionFilename("hello.swift"))
-  lazy val currentSwiftDefaultKind = "swift:3"
+  lazy val currentSwiftDefaultKind = "swift:3.1.1"
 
   behavior of "Swift runtime"
