+++ tests/src/test/scala/system/basic/WskRestBasicSwift311Tests.scala
@@ -25,5 +25,4 @@ import org.scalatest.junit.JUnitRunner
 @RunWith(classOf[JUnitRunner])
 class WskRestBasicSwift311Tests extends WskBasicSwift3Tests {
   override val wsk: common.rest.WskRest = new WskRest
-  override lazy val currentSwiftDefaultKind = "swift:3.1.1"
 }
